backend/requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
pydantic==2.10.6
sqlalchemy==2.0.37
python-dotenv==1.0.1